# Conflicts:
#	app/src/main/java/com/example/sandboxbank/App/ui/designkit/mode/color/ColorPalette.kt
#	app/src/main/java/com/example/sandboxbank/App/ui/mainscreen/ui/NavGraph.kt
#	app/src/main/java/com/example/sandboxbank/main/ui/HostActivity.kt

+++ app/src/main/java/com/example/sandboxbank/App/core/deposit/domain/model/FinancialItem.kt
@@ -1,12 +1,12 @@
 package com.example.sandboxbank.App.core.deposit.domain.model
 
 sealed class FinancialItem {
-    abstract val id: Int
+    abstract val id: Long
     abstract val type: String
     abstract val openDate: Long
     abstract val percentRate: Double
-    abstract val percentType: Int
-    abstract val period: Int
-    abstract val balance: Double
+    abstract val percentType: Long
+    abstract val period: Long
+    abstract val balance: Long
     abstract val name: String
 }
